+++ rocketgram/api/__init__.py
@@ -14,9 +14,11 @@ from .bot_command import BotCommand
 from .callback_query import CallbackQuery
 from .chat import Chat
 from .chat_action_type import ChatActionType
+from .chat_invite_link import ChatInviteLink
 from .chat_location import ChatLocation
 from .chat_member import ChatMember
 from .chat_member_status_type import ChatMemberStatusType
+from .chat_member_updated import ChatMemberUpdated
 from .chat_permissions import ChatPermissions
 from .chat_photo import ChatPhoto
 from .chat_type import ChatType